Rename `Bag` too `bags`.
