+++ src/components/bags/index.js
